feat(aoc-24): 21 solution
Today's problem was really hard to solve. Since the start that I tried
using some sort of function composition, in a recursive way, but it was
not getting me anywhere.
Ended up searching for an explanation of the problem in Reddit as I was
not seeing the end of the tunnel...